// src/app/core/components/notification-container/models/notification-type.enum.ts
export enum NotificationTypeEnum {
    JOIN = "JOIN",
    REQUEST_APPROVED = "REQUEST_APPROVED",
    REQUEST_REJECTED = "REQUEST_REJECTED",
    TRAVEL_CANCELED = "TRAVEL_CANCELED",
    APPLIED_TRAVEL_CANCELED = "APPLIED_TRAVEL_CANCELED",
    PASSENGER_LEFT = "PASSENGER_LEFT",
    PASSENGER_KICKED = "PASSENGER_KICKED",
    DRIVER_TRAVEL_TODAY = "DRIVER_TRAVEL_TODAY",
    PASSENGER_TRAVEL_TODAY = "PASSENGER_TRAVEL_TODAY"
}